Edit to make age a float so it can be less than 1

+++ create_db.py
@@ -8,6 +8,6 @@ connect("mongodb://bme590:hello12345@ds157818.mlab.com:57818/hr")
 class User(MongoModel):
     patient_id = fields.IntegerField(primary_key=True)
     attending_email = fields.EmailField()
-    user_age = fields.IntegerField()
+    user_age = fields.FloatField()
     heart_rate = fields.ListField(field=fields.BigIntegerField())
     time_stamp = fields.ListField(field=fields.DateTimeField())
